% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcca.R
\name{gcca}
\alias{gcca}
\title{Generalized canonical correlation}
\usage{
gcca(
  x,
  nfac = 2,
  scale = TRUE,
  pval = TRUE,
  scores = FALSE,
  method = "solve",
  lambda,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{x}{list of matrices. Each matrix should have the ids in the
rownames. Missing is not allowed (see details)}

\item{nfac}{...}

\item{scale}{...}

\item{pval}{should p-values of correlation between variables and shared canonical variates be computed? Default is TRUE.}

\item{scores}{should canonical variables be computed for each table?
Default is FALSE. See details}

\item{method}{...}

\item{lambda}{...}

\item{mc.cores}{...}
}
\value{
a list consisting of
  \item{Y}{canonical components for the shared space}
  \item{corsY}{correlation between variables and shared canonical components}
  \item{scores}{canonical componets for each table}
  \item{p.values}{p-values of correlation between variables and shared canonical components}
  \item{AVE}{indicators of model quality based on the Average Variance Explained (AVE): AVE for each table, AVE_outer (average accross tables), AVE_inner(for the shared component).}
}
\description{
Generalized canonical correlation
}
\details{
eigen with Rfast hd.eigen select number of components speed up process (RSpectra?)
rownames matrices ... si hay distintos ordena - si no, no.
missings ...  dos opciones#'
scores .. si queremos correlation with each table!
inversa ...
"solve" para n>>p (pero lento) (solve function)
"penalized" penalized adding lambda*I (rfunctions cgls). Requires lambda
"geninv" n<<p generalized inverse (rfunctions geninv)
"ginv" generalized inverse (MASS ginv)
}
\examples{
see vignette

}
