% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_AVE}
\alias{get_AVE}
\title{Compute Matrix Standard Deviation and Mean in HDF5}
\usage{
get_AVE(filename, group, datasets, wsize = NULL)
}
\description{
Computes standard deviation and/or mean statistics for a matrix stored in
HDF5 format, with support for row-wise or column-wise computations.
}
