% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDimNames_hdf5.R
\name{getColNames_hdf5}
\alias{getColNames_hdf5}
\title{Get hdf5 column names}
\usage{
getColNames_hdf5(filename, group, datasetname, sort = FALSE)
}
\arguments{
\item{filename}{character array indicating the name of the file to create}

\item{group}{path to dataset, character array indicating the complete route to the dataset}

\item{daasetname}{character array with dataset name}
}
\value{
List of tables
}
\description{
Get hdf5 column names
}
