% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcca_hdf5.R
\name{mgcca_hdf5}
\alias{mgcca_hdf5}
\title{Generalized canonical correlation with missing individuals for big data}
\usage{
mgcca_hdf5(
  x,
  datanames,
  nfac = 2,
  scale = TRUE,
  pval = TRUE,
  scores = FALSE,
  method = "solve",
  lambda,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{x}{string for hdf5 file where data to process will be under MGCCA_IN group .
Results will be stored in the same data file under MGCCA_OUT group. Missing is not allowed  (see details).}

\item{datanames}{string array with datasetnames to use with mcgga}

\item{nfac}{...}

\item{scale}{...}

\item{pval}{should p-values of correlation between variables and shared canonical variates be computed? Default is TRUE.}

\item{scores}{should canonical variables be computed for each table?
Default is FALSE. See details}

\item{method}{...}

\item{lambda}{...}

\item{mc.cores}{...}
}
\value{
a list consisting of
  \item{Y}{canonical components for the shared space}
  \item{corsY}{correlation between variables and shared canonical components}
  \item{scores}{canonical componets for each table}
  \item{p.values}{p-values of correlation between variables and shared canonical components}
  \item{AVE}{indicators of model quality based on the Average Variance Explained (AVE): AVE for each table, AVE_outer (average accross tables), AVE_inner(for the shared component).}
}
\description{
Generalized canonical correlation with missing individuals for big data
}
\details{
NOT WORKING
}
\examples{
see vignette

}
