% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgcca_svd_rcpp}
\alias{mgcca_svd_rcpp}
\title{MGCCA dual SVD stage: M_j = U diag(w) U' from SVD(X_j)}
\usage{
mgcca_svd_rcpp(
  filename,
  tmp_group,
  datasets,
  inv_method,
  lambda = NULL,
  svd_method = "full",
  threads = NULL
)
}
\description{
MGCCA dual SVD stage: M_j = U diag(w) U' from SVD(X_j)
}
\keyword{internal}
