% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgcca_XKX_rcpp}
\alias{mgcca_XKX_rcpp}
\title{MGCCA XKX stage (Mgram=X'X, inverse, Mi=X xkx X') over HDF5}
\usage{
mgcca_XKX_rcpp(
  filename,
  tmp_group,
  datasets,
  inv_method,
  lambda = NULL,
  threads = NULL
)
}
\description{
MGCCA XKX stage (Mgram=X'X, inverse, Mi=X xkx X') over HDF5
}
\keyword{internal}
