% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reliability_proj_overlap}
\alias{reliability_proj_overlap}
\title{K3b projector overlap trPP = tr(Pa Pb)/L}
\usage{
reliability_proj_overlap(Pa, Pb, L)
}
\value{
A single number: \eqn{tr(P_a P_b)/L}, the projector overlap of the
  two subspaces.
}
\description{
K3b projector overlap trPP = tr(Pa Pb)/L
}
\keyword{internal}
