% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCoords.R
\name{genCoords}
\alias{genCoords}
\title{Subset a MultiAssayExperiment object by specific genomic coordinates}
\usage{
genCoords(multiassayexperiment, gen.coords, meth.index, col.name)
}
\arguments{
\item{gen.coords}{must be a GRanges object}

\item{meth.index}{is the index where the methylation data (or a SummarizedExperiment non Ranged) is located inside the mae.}

\item{col.name}{column name}
}
\description{
Subset a MultiAssayExperiment object by specific genomic coordinates
}
