% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignif.R
\name{getSignif}
\alias{getSignif}
\title{Generalized canonical correlation with missing individuals}
\usage{
getSignif(x, df = NA, pval.cut = 0.05, ...)
}
\arguments{
\item{x}{an object of class 'mgcca'}
}
\description{
Generalized canonical correlation with missing individuals
}
