% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTables.R
\name{getTables}
\alias{getTables}
\title{Splits a MultiAssayExperiment in a list of tables}
\usage{
getTables(multiassayexperiment)
}
\arguments{
\item{multiassayexperiment}{MultiAssayExperiment}
}
\value{
List of tables
}
\description{
Splits a MultiAssayExperiment in a list of tables
}
