% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTables_hdf5.R
\name{getTables_hdf5}
\alias{getTables_hdf5}
\title{Splits a MultiAssayExperiment in a list of tables}
\usage{
getTables_hdf5(
  multiassayexperiment,
  filename,
  overwriteFile = FALSE,
  overwriteDataset = FALSE
)
}
\arguments{
\item{multiassayexperiment}{MultiAssayExperiment}

\item{filename}{string with file name}

\item{overwriteFile}{boolean, if true , file is overwritten}

\item{overwriteDataset}{boolean, if true , dataset is overwritten}
}
\value{
List of tables
}
\description{
Splits a MultiAssayExperiment in a list of tables
}
