% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Imputes each assay matrix data from a MultiAssayExperiment object}
\usage{
impute(
  multiassayexperiment,
  method,
  remove.col = FALSE,
  impute.zero = FALSE,
  rowmax = 0.5,
  colmax = 0.8,
  ...
)
}
\arguments{
\item{method}{...}

\item{remove.col}{remove cols}

\item{remove.row}{remove rows}
}
\description{
Imputes each assay matrix data from a MultiAssayExperiment object
}
