% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{matrix.chr2num}
\alias{matrix.chr2num}
\title{Formats a character matrix into a numeric matrix (used for the methylation assay)}
\usage{
matrix.chr2num(listMAE, index)
}
\arguments{
\item{index}{indicates in which position from the list is the matrix}
}
\description{
Formats a character matrix into a numeric matrix (used for the methylation assay)
}
