% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgccaConfig.R
\docType{package}
\name{mgcca-package}
\alias{mgcca-package}
\title{mgcca: Generalized Canonical Correlation Analysis with missing individuals}
\description{
Generalized Canonical Correlation Analysis (van de Velden and Takane, 2012)
for multi-block omics integration, running its numerical core in C++ on HDF5
files via the BigDataStatMeth API. See \code{\link{mgcca}} to get started.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://brge.isglobal.org}
}

}
\author{
\strong{Maintainer}: Dolors Pelegri-Siso \email{dolors.pelegri@isglobal.org} (\href{https://orcid.org/0000-0002-5993-3003}{ORCID})

Authors:
\itemize{
  \item Isaac Subirana
  \item Juan R. Gonzalez \email{juanr.gonzalez@isglobal.org} (\href{https://orcid.org/0000-0003-3267-2146}{ORCID})
}

}
\keyword{internal}
