% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcca.R
\name{mgcca}
\alias{mgcca}
\title{Generalized canonical correlation with missing individuals}
\usage{
mgcca(x, nfac = 2, scale = TRUE, mc.cores = 1, method = "Rfast",
  inv = "solve", ...)
}
\arguments{
\item{x}{list of matrices. Each matrix should have the ids in the
rownames. Missing is not allowed (see details)}

\item{nfac}{...}

\item{scale}{...}

\item{mc.cores}{...}

\item{method}{...}

\item{inv}{...}
}
\value{
A list of three components with ...
}
\description{
Generalized canonical correlation with missing individuals
}
\details{
eigen with Rfast hd.eigen select number of components speed up process
rownames matrices ...  avisar que se reordenan y que pueden haber más (todos)
missings ...  dos opciones#'
inversa ...
"solve" para n>>p (pero lento) (solve function)
"fast" para n>>p con n grande (Rfast spdinv)
"geninv" n<<p generalized inverse (rfunctions geninv)
"penalized" penalized adding lambda*I (rfunctions cgls)
"ginv" generalized inverse (MASS ginv)
}
\examples{
see vignette
}
