% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgcca_cor_ave_rcpp}
\alias{mgcca_cor_ave_rcpp}
\title{MGCCA corsY / p-values / AVE stage over HDF5}
\usage{
mgcca_cor_ave_rcpp(
  filename,
  tmp_group,
  datasets,
  nfac,
  final_group = "FINAL_RESULTS"
)
}
\description{
MGCCA corsY / p-values / AVE stage over HDF5
}
\keyword{internal}
