% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgcca_eigen_rcpp}
\alias{mgcca_eigen_rcpp}
\title{MGCCA eigen stage (Ksum, MKsum05, eigen, Y) over HDF5}
\usage{
mgcca_eigen_rcpp(
  filename,
  tmp_group,
  datasets,
  nfac,
  final_group = "FINAL_RESULTS",
  threads = NULL
)
}
\description{
MGCCA eigen stage (Ksum, MKsum05, eigen, Y) over HDF5
}
\keyword{internal}
