% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgcca_getK_rcpp}
\alias{mgcca_getK_rcpp}
\title{MGCCA getK stage (build padded X and indicator diagonal K) over HDF5}
\usage{
mgcca_getK_rcpp(filename, in_group, datasets, out_group = "MGCCA_TMP")
}
\description{
MGCCA getK stage (build padded X and indicator diagonal K) over HDF5
}
\keyword{internal}
