% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgcca_scores_rcpp}
\alias{mgcca_scores_rcpp}
\title{MGCCA scores stage (A=B_j, weights, scores) over HDF5}
\usage{
mgcca_scores_rcpp(
  filename,
  tmp_group,
  datasets,
  nfac,
  final_group = "FINAL_RESULTS"
)
}
\description{
MGCCA scores stage (A=B_j, weights, scores) over HDF5
}
\keyword{internal}
