% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInds.R
\name{plotInds}
\alias{plotInds}
\title{Plot variables correlations with shared canonical components}
\usage{
plotInds(
  x,
  group,
  ax1 = 1,
  ax2 = 2,
  col.list,
  print.labels = FALSE,
  cex.label = 0.8,
  pos.leg = "bottomright",
  main = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class 'mgcca'}
}
\description{
Plot variables correlations with shared canonical components
}
