% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotScree}
\alias{plotScree}
\title{Scree plot of the shared-component eigenvalues}
\usage{
plotScree(x, title = "Eigenvalues")
}
\arguments{
\item{x}{an \code{mgcca} object.}

\item{title}{plot title. Default \code{"Eigenvalues"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Bar plot of the eigenvalues (inner-model AVE) associated with
  each shared component.
}
\seealso{
\code{\link{plotAVE}}
}
