% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVars.R
\name{plotVars}
\alias{plotVars}
\title{Plot variables correlations with shared canonical components}
\usage{
plotVars(x, var = NA, axes = 1:2, var.col = "red", var.lab = FALSE,
  bg.var.col = "gray", nlab = 0, df = NA, layout = NA, ...)
}
\arguments{
\item{x}{an object of class 'mgcca'}
}
\description{
Plot variables correlations with shared canonical components
}
