% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMAE.R
\name{print.listMAE}
\alias{print.listMAE}
\title{Print method for 'listMAE' objects}
\usage{
\method{print}{listMAE}(listMAE)
}
\arguments{
\item{listMAE}{Object of class 'listMAE'.}

\item{...}{Further arguments (ignored).}
}
\description{
Compact print for a MultiAssayExperiment-like list.
}
