% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcca_methods.R
\name{print.mgcca}
\alias{print.mgcca}
\title{Print an mgcca object}
\usage{
\method{print}{mgcca}(x, ...)
}
\arguments{
\item{x}{an \code{mgcca} object.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an mgcca object
}
